+++ src/pages/AboutMe/AboutMe.scss
@@ -1,7 +1,7 @@
 @import '../../scss/index.scss';
 
 .aboutMe {
-    background-image: url('../../assets/imgs/backgrounds/about.jpg');
+    background-image: url('../../assets/imgs/backgrounds/mac.jpg');
     // height: 95.4vh;
     &_container{
         // height: 100%;
